transfer: remove file transfer size limitation
File transfer is no longer limited to a file less than 16 KiB in size.
The limit of a single file remains, though.

Provides a more complete 'sendFile' implementation that can handle a
file of any size.

While the sending peer is no longer limited in the size of the file that
can be sent, the receiving peer may be limited in the size of the file
that can be received. Since the whole file is being buffered on the
receiving side, large files may cause a slowdown due to memory usage.

What happens when the memory capacity is reached, does the swap get
used? To be tested.