server: refactor, change ID flow and future-proof
Previously, the server would generate a session ID and give it to the
client. The client would then try to claim it when opening a socket. If
it was not unique, the server would tell the client to reload (getting a
new session ID) and try again. That worked, but could send the client
into a reload frenzy attempting to get a unique session ID, which is
quite jarring, especially if it persists because the session ID reserve
is (nearly) exhausted.

Now, the server guarantees that the session ID given to the client will
be unique. If the server cannot generate a unique session ID after
repeated attempts (currently 100), it will respond with a more
appropriate 503 Service Unavailable. This change required an additional
security mechanism using unique tokens to prevent session ID hijacking.

This new session ID flow opens the possibility for:
 * having multiple simultaneous transfers (sending and receiving), by
   allowing ownership of multiple session IDs, and
 * future use with a headless client.